% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{check_sdm_paths}
\alias{check_sdm_paths}
\title{Check if Paths are Valid}
\usage{
check_sdm_paths(paths = NULL)
}
\arguments{
\item{paths}{Optional list of paths to check. If NULL, checks currently set paths.}
}
\value{
Logical indicating if all paths are valid
}
\description{
Check if all specified data paths exist.
}
