% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_sdm_elevation}
\alias{download_sdm_elevation}
\title{Download elevation data}
\usage{
download_sdm_elevation(data_dir = NULL, force = FALSE)
}
\arguments{
\item{data_dir}{Directory to download data to. If NULL, uses the default.}

\item{force}{Logical. If TRUE, re-download even if files exist.}
}
\value{
Invisibly returns TRUE if successful, FALSE otherwise.
}
\description{
Download elevation data
}
\examples{
\dontrun{
download_sdm_elevation()
}
}
