% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{set_sdm_paths}
\alias{set_sdm_paths}
\title{Set Custom Data Paths for sistaR.sdm}
\usage{
set_sdm_paths(paths)
}
\arguments{
\item{paths}{List containing paths to environmental data. Can include:
\itemize{
\item{climate}{ - Directory containing climate .tif files}
\item{elevation}{ - Path to elevation .tif file}
\item{soil_raster}{ - Path to HWSD2 .tif raster file}
\item{soil_db}{ - Path to HWSD2 .sqlite database file}
\item{srad}{ - Directory containing solar radiation .tif files}
\item{wind}{ - Directory containing wind speed .tif files}
\item{vapr}{ - Directory containing vapor pressure .tif files}
}}
}
\value{
Invisibly returns the previously set paths
}
\description{
Set or update custom paths to environmental data files.
}
\examples{
\dontrun{
set_sdm_paths(list(
  climate = "D:/SDM_data/climate/wc2.1_10m",
  elevation = "D:/SDM_data/elevation/wc2.1_10m_elev.tif",
  soil_raster = "D:/SDM_data/soil/HWSD2_RASTER/HWSD2.tif",
  soil_db = "D:/SDM_data/soil/HWSD2_DB/HWSD2.sqlite"
))
}
}
